A3.BasitHavaDurumuPortali/requirements.txt
selenium>=4.0.0
Pillow>=9.0.0
webdriver-manager>=3.8.0
